Add files via upload
In this short notebook, I've implemented a hybrid (classical+quantum) realisation of the Durr-Hoyer Algorithm. This project was done within the course "Quantum and Classical Machine Learning". The primary goal was to study NN classification, which uses the Durr-Hoyer algorithm (a variant of Grover search) for finding the closest point ( https://arxiv.org/abs/1401.2142), but instead of the euclidean distance, which was used in [1] to choose the Hamming distance
[2] (https://arxiv.org/pdf/2103.04253.pdf) due to its simplicity.